Parser can now bound headers as a ReadOnlyBuffer
This functionality is required to implement the Select/Operate checking
functionality easily. It's very light-weight and has minimal impact for
other object types.

+++ cpp/opendnp3/src/opendnp3/Iterable.h
@@ -44,7 +44,7 @@ class Iterable
 		
 		virtual Iterator<T> GetIterator() const = 0;		
 
-		virtual size_t Size() const = 0;
+		virtual uint32_t Size() const = 0;
 };
